+++ elements/files/translations/de.php
@@ -1,4 +1,4 @@
 <?php
 
 l::set('panelBar.element.files.label', 'Dateien');
-l::set('panelBar.element.files.all', 'Alle');
+l::set('panelBar.element.files.all', 'Alle Dateien');