+++ elements/files/translations/en.php
@@ -1,4 +1,4 @@
 <?php
 
 l::set('panelBar.element.files.label', 'Files');
-l::set('panelBar.element.files.all', 'All');
+l::set('panelBar.element.files.all', 'All Files');