+++ elements/images/translations/de.php
@@ -1,4 +1,4 @@
 <?php
 
 l::set('panelBar.element.images.label', 'Bilder');
-l::set('panelBar.element.images.all', 'Alle');
+l::set('panelBar.element.images.all', 'Alle Bilder');