+++ elements/images/translations/en.php
@@ -1,4 +1,4 @@
 <?php
 
 l::set('panelBar.element.images.label', 'Images');
-l::set('panelBar.element.images.all', 'All');
+l::set('panelBar.element.images.all', 'All Images');