+++ elements/system/translations/en.php
@@ -1,0 +1,9 @@
+<?php
+
+l::set('panelBar.element.system.label', 'Site');
+l::set('panelBar.element.system.pages', 'Pages');
+l::set('panelBar.element.system.files', 'Files');
+l::set('panelBar.element.system.templates', 'Templates');
+l::set('panelBar.element.system.blueprints', 'Blueprints');
+l::set('panelBar.element.system.controllers', 'Controllers');
+l::set('panelBar.element.system.models', 'Models');