+++ elements/user/user.php
@@ -17,8 +17,7 @@ class UserElement extends Element {
         'label' => $user,
         'icon'  => 'user',
         'url'   => $user->url('edit'),
-        'title' => 'User: ' . $user,
-        'right' => true,
+        'title' => $this->l('user' , ['user' => $user->username()]),
       ]);
     }
   }