+++ elements/visibility/translations/de.php
@@ -6,5 +6,5 @@ l::set('panelBar.element.visibility.invisible',  'Unsichtbar');
 l::set('panelBar.element.visibility.makevisible',  'Anzeigen');
 l::set('panelBar.element.visibility.makeinvisible',  'Verstecken');
 
-l::set('panelBar.element.visibility.setbefore',  'Einfügen vor');
-l::set('panelBar.element.visibility.setafter',  'Einfügen nach');
+l::set('panelBar.element.visibility.setbefore',  'Einfügen vor {page}');
+l::set('panelBar.element.visibility.setafter',  'Einfügen nach {page}');