+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Changelog
 
+## [2.2.0](https://github.com/distantnative/panel-bar/releases/tag/2.2.0) (2017-04-9)
+:sparkles: New element: Google PageSpeed API
+:sparkles: Set element float via panel widget
+:lipstick: Panel widget: now all pretty
+:balloon: Translations now follow the panel user's language  
+:balloon: Translations now get passed {$variables} to be more flexible 
+:poop: Dropped RTL support for the moment   
+
 ## [2.1.2](https://github.com/distantnative/panel-bar/releases/tag/2.1.2) (2017-03-28)
-:bug: CLI installing to wrong destination
-:bug: active elements not considering default element set
+:bug: CLI installing to wrong destination  
+:bug: active elements not considering default element set  
 
 ## [2.1.1](https://github.com/distantnative/panel-bar/releases/tag/2.1.1) (2017-03-26)
 :bug: Visibility element: Unhiding numbered pages works again